% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portals.R
\name{immor_portal}
\alias{immor_portal}
\title{Get a configured portal by name}
\usage{
immor_portal(name)
}
\arguments{
\item{name}{Character name of the portal (e.g.,
\code{"flatfox"}).}
}
\value{
An \code{immor_portal} object.
}
\description{
Get a configured portal by name
}
\examples{
immor_portal("flatfox")
}
