% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portals.R
\name{immor_portals}
\alias{immor_portals}
\title{List available portals}
\usage{
immor_portals()
}
\value{
A named list of portal constructor functions.
}
\description{
Returns a named list of all portal constructors
registered in immor.
}
\examples{
immor_portals()
names(immor_portals())
}
