% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{immor_query}
\alias{immor_query}
\title{Build a search query for immor}
\usage{
immor_query()
}
\value{
An S3 object of class \code{immor_query}.
}
\description{
Constructs a query object used to search across Swiss
real estate portals.
}
\examples{
immor_query()
}
