% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal-flatfox.R
\name{portal_flatfox}
\alias{portal_flatfox}
\title{Create a flatfox portal}
\usage{
portal_flatfox()
}
\value{
An \code{immor_portal} object for flatfox.
}
\description{
Constructs a portal object for
\href{https://flatfox.ch}{flatfox.ch}, which provides a
public REST API.
}
\examples{
portal_flatfox()
}
